minor fixes

+++ Assets/DTWorldz/Scripts/ScriptableObjects/Items/Equipments/BaseEquipment.cs
@@ -5,7 +5,7 @@ namespace DTWorldz.ScriptableObjects.Items.Equipments
 {
     public abstract class BaseEquipment : BaseItem
     {
-        public float MinStrength;
-        public float MinDexterity;
+        public int MinStrength;
+        public int MinDexterity;
     }
 }
